refactor: pageSelector 경로 파일 분리
