+++ src/pages/Profile.js
@@ -4,15 +4,20 @@ import styled from "styled-components";
 import Navbar from "../components/global/Navbar";
 import PageSelector from "../components/global/PageSelector";
 import UserProfile from "../components/user/Profile";
-import { temp_profile_pageselector, temp_user_profile } from "../consts/tempData";
+import { temp_user_profile } from "../consts/tempData";
 import Layout from "../layouts/Layout";
+import { temp_profile_pageselector } from "../consts/pageData";
 
 export default function Profile() {
   const navigate = useNavigate();
   const handleEditOpen = () => navigate(`edit`);
   const handlePrevPage = () => navigate(-1);
   return (
-    <Layout headerLeft={"prev"} headerRight={"setting"} onClickLeft={handlePrevPage}>
+    <Layout
+      headerLeft={"prev"}
+      headerRight={"setting"}
+      onClickLeft={handlePrevPage}
+    >
       <UserProfile user={temp_user_profile} isMyProfile={true}>
         <FaPen size={20} onClick={handleEditOpen} />
       </UserProfile>
