+++ src/pages/club/Club.js
@@ -1,11 +1,11 @@
 import { Outlet, useNavigate } from "react-router-dom";
 import styled from "styled-components";
 import PageSelector from "../../components/global/PageSelector";
-import { temp_club_pageselector } from "../../consts/tempData";
 import usePrevPage from "../../hooks/usePrevPage";
 import Layout from "../../layouts/Layout";
 import { useState } from "react";
 import ClubSettingDropDown from "../../components/club/ClubSettingDropDow";
+import { temp_club_pageselector } from "../../consts/pageData";
 
 export default function Club() {
   const { onPrevPage } = usePrevPage();
