+++ src/pages/clubSetting/ClubSetting.js
@@ -1,9 +1,9 @@
 import { Outlet } from "react-router-dom";
 import PageSelector from "../../components/global/PageSelector";
-import { temp_club_setting_page } from "../../consts/tempData";
 import Layout from "../../layouts/Layout";
 import styled from "styled-components";
 import usePrevPage from "../../hooks/usePrevPage";
+import { temp_club_setting_page } from "../../consts/pageData";
 
 export default function ClubSetting() {
   const { onPrevPage } = usePrevPage();
